+++ sources/packages/web/src/components/aest/students/assessment/ManualReassessment.vue
@@ -30,7 +30,7 @@ import { ModalDialog, useSnackBar } from "@/composables";
 import { ManualReassessmentAPIInDTO } from "@/services/http/dto";
 import TriggerReassessmentModal from "@/components/aest/students/modals/TriggerReassessmentModal.vue";
 import { ApplicationService } from "@/services/ApplicationService";
-import { Role, StudentAssessmentStatus } from "@/types";
+import { ApplicationStatus, Role, StudentAssessmentStatus } from "@/types";
 import CheckPermissionRole from "@/components/generic/CheckPermissionRole.vue";
 
 export default defineComponent({
@@ -62,7 +62,12 @@ export default defineComponent({
       openModalButtonDisabled.value =
         applicationAssessmentStatusDetails.originalAssessmentStatus !==
           StudentAssessmentStatus.Completed ||
-        applicationAssessmentStatusDetails.isApplicationArchived;
+        applicationAssessmentStatusDetails.isApplicationArchived ||
+        [
+          ApplicationStatus.Cancelled,
+          ApplicationStatus.Overwritten,
+          ApplicationStatus.Draft,
+        ].includes(applicationAssessmentStatusDetails.applicationStatus);
     });
 
     const openTriggerReassessmentModal = async () => {
