+++ sources/packages/web/src/services/http/dto/Assessment.dto.ts
@@ -83,3 +83,7 @@ export interface AwardDetailsAPIOutDTO {
    */
   finalAward?: Record<string, string | number>;
 }
+
+export interface ManualReassessmentAPIInDTO {
+  note: string;
+}
